:construction: Make some changes to testing storyboard app
